Add function to read edgelist, in parallel

+++ minigraph/portability/sys_types.h
@@ -4,7 +4,7 @@
 
 #include <gflags/gflags.h>
 
-using gid_t = size_t;
+using gid_t = unsigned;
 using vid_t = size_t;
 using vdata_t = float;
 using edata_t = size_t;
